% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{get_giphy_api_key}
\alias{get_giphy_api_key}
\alias{set_giphy_api_key}
\title{Get and set giphy api key}
\usage{
get_giphy_api_key()

set_giphy_api_key(key)
}
\arguments{
\item{key}{An api key string.}
}
\description{
Get and set the giphy api key. See https://github.com/Giphy/GiphyAPI for
more details.
}
