% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{praise_with_giphy}
\alias{praise_with_giphy}
\title{Get a praise template with a gif}
\usage{
praise_with_giphy(template = "You are ${adjective}!", rating = "g",
  size = "downsized_medium")
}
\arguments{
\item{template}{The template string. See \link[praise:praise]{praise::praise} for more details.}

\item{rating}{Giphy's MPAA rating filter. Available options are: "y", "g", "pg", "pg-13" and "r".}

\item{size}{Giphy's rendition class. Some available options are "original" (original size),
"downsized_medium" (file size under 5mb), and others. See the Giphy API for
more details.}
}
\description{
Pairs \link[praise:praise]{praise::praise} with a gif using giphy's search. Returns a \code{praise_giphy}
object, which prints the html needed to embed the document (and prints in a knitr
document).
}
